+++ migrations/20210408200416_create_food_table.js
@@ -1,7 +1,7 @@
 exports.up = function (knex) {
     return knex.schema.createTable("food", (table) => {
       table.increments("id");
-      table.bigInteger('result_food_total');
+      table.decimal('result_food_total');
       table.timestamp("time_input").defaultTo(knex.fn.now());
       table.integer("user_id").unsigned().references("users.id");
     });
