+++ migrations/20210410004629_transportation_table.js
@@ -1,7 +1,7 @@
 exports.up = function (knex) {
     return knex.schema.createTable('transportation', (table) => {
       table.increments('id');
-      table.bigInteger('result_transport_total');
+      table.decimal('result_transport_total');
       table.timestamp('time_input').defaultTo(knex.fn.now());
       table.integer('user_id').unsigned().references('users.id');
     });
